feat(MenuVoteService): 투표현황을 조회하고 퍼센테이지를 계산하는 서비스 구현
라멘샵 컨트롤러에서 투표기능 분리

+++ src/main/java/com/raota/domain/ramenShop/controller/response/VotingStatusResponse.java
@@ -4,7 +4,7 @@ import com.raota.domain.ramenShop.dto.VoteResultsDto;
 import java.util.List;
 
 public record VotingStatusResponse (
-        int total_votes,
+        long total_votes,
         List<VoteResultsDto> vote_results
 ){
 }
